added the security fetures
