+++ ios/Runner/AppDelegate.swift
@@ -10,4 +10,16 @@ import Flutter
     GeneratedPluginRegistrant.register(with: self)
     return super.application(application, didFinishLaunchingWithOptions: launchOptions)
   }
+  
+  override func applicationWillResignActive(
+    _ application: UIApplication
+  ) {
+    window?.rootViewController?.view.endEditing(true)
+    self.window.isHidden = true;
+  }
+  override func applicationDidBecomeActive(
+    _ application: UIApplication
+  ) {
+    self.window.isHidden = false;
+  }
 }
